pyproject.toml
[tool.poetry]
name = "chatapp-api"
version = "0.0.1"
description = "API for real time communication with friends."
authors = ["Togrul Asadov <adsads00123@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
aioredis = "^2.0.1"
alembic = "1.8.1"
anyio = "3.6.1"
asyncpg = "^0.27.0"
attrs = "22.1.0"
bcrypt = "4.0.0"
broadcaster = { extras = ["redis"], version = "^0.2.0" }
cffi = "1.15.1"
certifi = "2022.12.7"
charset-normalizer = "2.1.1"
click = "8.1.3"
colorama = "0.4.5"
cryptography = "39.0.1"
dnspython = "2.2.1"
ecdsa = "0.18.0"
email-validator = "1.2.1"
fastapi = "0.88.0"
greenlet = "1.1.3"
h11 = "0.13.0"
httptools = "0.4.0"
idna = "3.3"
iniconfig = "1.1.1"
mako = "1.2.2"
markupsafe = "2.1.1"
mccabe = "0.7.0"
packaging = "21.3"
passlib = "1.7.4"
pillow = "9.4.0"
pluggy = "1.0.0"
py = "1.11.0"
pyasn1 = "0.4.8"
pycodestyle = "2.9.1"
pycparser = "2.21"
pydantic = "1.10.1"
pyflakes = "2.5.0"
pyparsing = "3.0.9"
python = "^3.10"
python-dotenv = "0.21.0"
python-jose = "3.3.0"
python-multipart = "0.0.5"
pyyaml = "6.0"
requests = "2.28.1"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = { extras = ["asyncio", "mypy"], version = "2.0.9" }
starlette = "0.22.0"
tomli = "2.0.1"
urllib3 = "1.26.12"
uvicorn = "0.18.3"
watchfiles = "0.16.1"
websockets = "10.3"

[tool.poetry.group.dev.dependencies]
asyncpg-stubs = "^0.27.0"
bandit = { extras = ["toml"], version = "^1.7.4" }
black = "^22.12.0"
coverage = "6.4.4"
flake8 = "5.0.4"
flake8-isort = "^6.0.0"
httpx = "^0.23.3"
isort = "^5.11.4"
mypy = "^0.991"
pre-commit = "^2.21.0"
pylint = { extras = ["spelling"], version = "^2.15.9" }
pytest = "7.1.3"
pytest-asyncio = "^0.20.3"
pytest-cov = "3.0.0"
typing-extensions = "4.3.0"
types-passlib = "^1.7.7.7"
types-python-jose = "^3.3.4.3"
types-pillow = "^9.4.0.6"
flake8-bandit = "^4.1.1"
safety = "^2.3.5"
toml-sort = "^0.22.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
assert_used = { skips = ["*/test_*.py"] }

[tool.black]
line-length = 79
target-version = ["py310"]
extend-exclude = """
^/src/chatapp_api/alembic/versions
"""

[tool.isort]
profile = "black"
line_length = 79
skip = ["./src/chatapp_api/alembic/versions"]

[tool.mypy]
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
exclude = ["/versions/"]
explicit_package_bases = true


[tool.pylint]
ignore = "versions"

[tool.pytest.ini_options]
testpaths = "./tests"
python_files = "test_*.py"
asyncio_mode = "auto"
pythonpath = "."
